requirements.txt
Flask
tensorflow
Pillow
requests
numpy
werkzeug
gunicorn
scipy
